// my-app/src/app/services/documento/documento-interface.ts
export interface DocumentoInterface {
    id?: number;
    subject?: string;
    control?: string;
    type_id?: number;
    url?: string;
    expiration?: string;
    active?: string;
    conclution_id?: number;
    preference_id?: number;
    office_id?: number;
    user_id?: number;
    created_at?: string;
    updated_at?: string;

    state_id?: number;
    feedback?: string;
}